# Optimize the performance of the reduce_circuit function in specific situations