--- a/doc/source/tutorials/Circuit.ipynb
+++ b/doc/source/tutorials/Circuit.ipynb
@@ -291,6 +291,140 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
+    "### Circuit Reduction\n",
+    "Here we model a slightly more complex band-pass LC filter to demonstrate circuit reduction, with example values taken from [markimicrowave.com](https://markimicrowave.com/technical-resources/tools/lc-filter-design-tool/) :\n",
+    "\n",
+    "![low pass filter](figures/circuit_filter2.svg)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "freq = rf.Frequency(50, 200, 301, 'mhz')\n",
+    "tl_media = rf.media.DefinedGammaZ0(frequency=freq, z0=50)\n",
+    "gnd = rf.Circuit.Ground(frequency=freq,name='ground')\n",
+    "C1 = tl_media.capacitor(6.353e-12, name='C1')\n",
+    "L1 = tl_media.inductor(402.7e-9, name='L1')\n",
+    "C2 = tl_media.capacitor(61.08e-12, name='C2')\n",
+    "L2 = tl_media.inductor(13.63e-9, name='L2')\n",
+    "C3 = tl_media.capacitor(187.8e-12, name='C3')\n",
+    "L3 = tl_media.inductor(41.89e-9, name='L3')\n",
+    "C4 = tl_media.capacitor(6.353e-12, name='C4')\n",
+    "L4 = tl_media.inductor(402.27e-9, name='L4')\n",
+    "Port1 = rf.Circuit.Port(frequency=freq, name='PortIn')\n",
+    "Port2 = rf.Circuit.Port(frequency=freq, name='PortOut')\n",
+    "cnx = [\n",
+    "    [(Port1, 0), (C1, 0)],\n",
+    "    [(C1, 1), (L1, 0)],\n",
+    "    [(L1, 1), (C2, 0), (C3, 0), (C4, 0)],\n",
+    "    [(C2, 1), (L2, 0)],\n",
+    "    [(C3, 1), (L3, 0)],\n",
+    "    [(L2, 1), (L3, 1), (gnd, 0)],\n",
+    "    [(C4, 1), (L4, 0)],\n",
+    "    [(L4, 1), (Port2, 0)]\n",
+    "]\n",
+    "cir = rf.Circuit(cnx)\n",
+    "ntw = cir.network"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "ntw.plot_s_db(m=0, n=0, lw=2)\n",
+    "ntw.plot_s_db(m=1, n=0, lw=2)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "As in the previous example, we can use [Circuit.plot_graph()](../api/circuit.html#representing-a-circuit) method to represent the connexion graphically."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "cir.plot_graph(network_labels=True, network_fontsize=15,\n",
+    "               port_labels=True, port_fontsize=15,\n",
+    "              edge_labels=True, edge_fontsize=10)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "When focusing solely on `Ports` and disregarding internal `Networks`, we could utilize [rf.circuit_reducte()](../api/circuit.html#circuit-reduction) to streamline the circuit. This method will automatically series-connect adjacent networks through a `rf.connect()` approach, retaining only `Ports` and connections containing more than 2 `Networks` for circuit analysis."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "reduced_cnx = rf.reduce_circuit(cnx)\n",
+    "reduced_cir = rf.Circuit(reduced_cnx)\n",
+    "reduced_cir.plot_graph(network_labels=True, network_fontsize=15,\n",
+    "               port_labels=True, port_fontsize=15,\n",
+    "              edge_labels=True, edge_fontsize=10)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Additionally, optional parameters can influence the behavior of circuit reduction. For instance, `split_ground=True` indicates splitting the shared `Ground` in the circuit:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "fully_reduced_cnx = rf.reduce_circuit(cnx, split_ground=True)\n",
+    "fully_reduced_cir = rf.Circuit(fully_reduced_cnx)\n",
+    "\n",
+    "fully_reduced_cir.plot_graph(network_labels=True, network_fontsize=15,\n",
+    "               port_labels=True, port_fontsize=15,\n",
+    "              edge_labels=True, edge_fontsize=10)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Finally, the Circuit constructor offers an option for automatic circuit simplification. Setting `auto_reduce=True` enables the creation of a significantly streamlined `Circuit` object, enhancing efficiency for Port's S-parameter calculations."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import numpy as np\n",
+    "fully_reduced_ntw = fully_reduced_cir.network\n",
+    "auto_reduced_cir = rf.Circuit(cnx, auto_reduce=True)\n",
+    "auto_reduced_ntw = auto_reduced_cir.network\n",
+    "\n",
+    "print(f\"Manually reduced circuit has the same s-parameters as the original circuit: {np.allclose(ntw.s, fully_reduced_ntw.s)}\")\n",
+    "print(f\"Auto reduced circuit has the same s-parameters as the original circuit: {np.allclose(ntw.s, auto_reduced_ntw.s)}\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
     "## References\n",
     "\n",
     "<div id=\"ref1\"></div>[1] Hallbjörner, P., 2003. Method for calculating the scattering matrix of arbitrary microwave networks giving both internal and external scattering. Microw. Opt. Technol. Lett. 38, 99–102. https://doi.org/10/d27t7m\n"