--- a/doc/source/tutorials/Circuit.ipynb
+++ b/doc/source/tutorials/Circuit.ipynb
@@ -412,10 +412,16 @@
    "metadata": {},
    "outputs": [],
    "source": [
+    "from timeit import timeit\n",
+    "\n",
     "import numpy as np\n",
+    "\n",
     "fully_reduced_ntw = fully_reduced_cir.network\n",
     "auto_reduced_cir = rf.Circuit(cnx, auto_reduce=True)\n",
     "auto_reduced_ntw = auto_reduced_cir.network\n",
+    "n_times = 10\n",
+    "print(f\"Unreduced Circuit calculate Network in {timeit(lambda: cir.network, number=n_times)/n_times:.4f}s\")\n",
+    "print(f\"Auto-reduced Circuit calculate Network in {timeit(lambda: auto_reduced_cir.network, number=n_times)/n_times:.4f}s\")\n",
     "\n",
     "print(f\"Manually reduced circuit has the same s-parameters as the original circuit: {np.allclose(ntw.s, fully_reduced_ntw.s)}\")\n",
     "print(f\"Auto reduced circuit has the same s-parameters as the original circuit: {np.allclose(ntw.s, auto_reduced_ntw.s)}\")"