--- a/doc/source/tutorials/Circuit.ipynb
+++ b/doc/source/tutorials/Circuit.ipynb
@@ -362,7 +362,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "When focusing solely on `Ports` and disregarding internal `Networks`, we could utilize [rf.circuit_reducte()](../api/circuit.html#circuit-reduction) to streamline the circuit. This method will automatically series-connect adjacent networks through a `rf.connect()` approach, retaining only `Ports` and connections containing more than 2 `Networks` for circuit analysis."
+    "When focusing solely on `Ports` and disregarding internal `Networks`, we could utilize [rf.reduce_circuit()](../api/circuit.html#circuit-reduction) to streamline the circuit. This method will automatically series-connect adjacent networks through a `rf.connect()` approach, retaining only `Ports` and connections containing more than 2 `Networks` for circuit analysis."
    ]
   },
   {
@@ -461,7 +461,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.8.13"
+   "version": "3.11.9"
   }
  },
  "nbformat": 4,