… doc/source/tutorials/figures/circuit_filter2.svg