This pull request improves the performance of the `reduce_circuit` method when connections are out of order and when multiple networks are connected to the same `Ground` network. Related to https://github.com/scikit-rf/scikit-rf/pull/1069

Benchmark code:
```python
import skrf as rf
from skrf import Circuit
import numpy as np
from time import time


def HFSS_cnxs(f_n: int):
    freq = rf.Frequency(start=1, stop=2, npoints=f_n, unit="GHz")

    hfss_ntwk = rf.Network(name="HFSS")
    hfss_ntwk.s=np.random.rand(f_n, 16, 16)+ 1.0j * np.random.rand(f_n, 16, 16)
    hfss_ntwk.frequency = freq

    line = rf.media.DefinedGammaZ0(frequency=freq, z0=50)

    Port1 = rf.Circuit.Port(frequency=freq, name="Port1")
    Port2 = rf.Circuit.Port(frequency=freq, name="Port2")
    Port3 = rf.Circuit.Port(frequency=freq, name="Port3")

    Ind1 = line.inductor(name="Ind1", L=3.85e-9)
    Ind2 = line.inductor(name="Ind2", L=5.94e-9)
    Ind3 = line.inductor(name="Ind3", L=3.14e-9)

    Cap1 = line.capacitor(name="Cap1", C=4.2e-12)
    Cap2 = line.capacitor(name="Cap2", C=0.35e-12)

    Gnd = rf.Circuit.Ground(frequency=freq, name="Gnd")

    ntwks = [rf.Network(name=f"ntwk{i}") for i in range(9)]
    for i in range(9):
        ntwks[i].frequency = freq
        ntwks[i].s = np.random.rand(f_n, 2, 2) + 1.0j * np.random.rand(
            f_n, 2, 2
        )
 
    connections = [
        [(hfss_ntwk, 10), (ntwks[5], 0)],
        [(hfss_ntwk, 11), (ntwks[6], 0)],
        [(hfss_ntwk, 12), (ntwks[7], 0)],
        [(hfss_ntwk, 0), (Ind1, 0), (Cap1, 0)],
        [(hfss_ntwk, 5), (ntwks[2], 0)],
        [(hfss_ntwk, 9), (ntwks[4], 0)],
        [
            (Ind1, 1),
            (Cap2, 1),
            (Ind3, 1),
            (ntwks[0], 1),
            (ntwks[1], 1),
            (ntwks[2], 1),
            (ntwks[4], 1),
            (ntwks[5], 1),
            (ntwks[6], 1),
            (hfss_ntwk, 14),
            (ntwks[8], 1),
            (Gnd, 0),
        ],
        [(hfss_ntwk, 7), (ntwks[3], 0)],
        [(hfss_ntwk, 8), (ntwks[3], 1)],
        [(Cap1, 1), (Port1, 0)],
        [(hfss_ntwk, 1), (Cap2, 0), (Ind2, 0)],
        [(Ind2, 1), (Port2, 0)],
        [(hfss_ntwk, 2), (Ind3, 0), (Port3, 0)],
        [(hfss_ntwk, 3), (ntwks[0], 0)],
        [(hfss_ntwk, 4), (ntwks[1], 0)],
        [(hfss_ntwk, 13), (ntwks[7], 1)],
        [(hfss_ntwk, 15), (ntwks[8], 0)],
    ]

    return connections


def complex_cnxs(f_n: int):
    freq = rf.Frequency(start=1, stop=2, npoints=f_n, unit="GHz")
    line = rf.media.DefinedGammaZ0(frequency=freq, z0=50)

    snp_n = 80
    Snp = [
        rf.Network(name=f"Snp{i}", s=np.random.rand(f_n, 2, 2)) for i in range(snp_n)
    ]
    for i in range(snp_n):
        Snp[i].frequency = freq

    # Gnds = [rf.Circuit.Ground(frequency=freq, name=f"Gnd{i}") for i in range(10)]
    Gnd = rf.Circuit.Ground(frequency=freq, name=f"Gnd")

    Input1 = rf.Circuit.Port(frequency=freq, name="Input1")
    Input2 = rf.Circuit.Port(frequency=freq, name="Input2")
    Ind = [line.inductor(name=f"Ind{i}", L=1e-9) for i in range(5)]
    Output1 = rf.Circuit.Port(frequency=freq, name="Output1")

    connections = [
        [
            (Snp[24], 0),
            (Snp[5], 0),
            (Snp[30], 0),
            (Snp[45], 0),
            (Ind[0], 1),
            (Snp[15], 0),
            (Snp[35], 0),
            (Gnd, 0),
        ],
        [(Snp[24], 1), (Snp[25], 0)],
        [(Snp[5], 1), (Snp[6], 0)],
        [(Input1, 0), (Snp[1], 0)],
        [(Snp[12], 0), (Snp[11], 1)],
        [(Snp[12], 1), (Snp[13], 0)],
        [(Snp[30], 1), (Snp[31], 0)],
        [(Snp[20], 0), (Snp[19], 1)],
        [(Snp[33], 0), (Snp[32], 1)],
        [(Snp[33], 1), (Snp[36], 1), (Snp[38], 0)],
        [(Input2, 0), (Snp[27], 0)],
        [(Snp[7], 0), (Snp[6], 1)],
        [(Snp[7], 1), (Snp[10], 1), (Snp[11], 0)],
        [(Snp[25], 1), (Snp[26], 0)],
        [(Snp[23], 0), (Snp[22], 1)],
        [(Snp[23], 1), (Output1, 0), (Snp[71], 1), (Ind[0], 0)],
        [(Snp[22], 0), (Snp[21], 1)],
        [(Snp[45], 1), (Snp[47], 0)],
        [(Snp[13], 1), (Snp[14], 0)],
        [(Snp[15], 1), (Snp[16], 0)],
        [(Snp[27], 1), (Snp[28], 0)],
        [(Snp[32], 0), (Snp[47], 1), (Snp[46], 1)],
        [(Snp[35], 1), (Snp[36], 0)],
        [(Snp[28], 1), (Snp[46], 0), (Snp[31], 1)],
        [(Snp[16], 1), (Snp[17], 0)],
        [(Snp[14], 1), (Snp[17], 1), (Snp[18], 0)],
        [(Snp[71], 0), (Snp[70], 1)],
        [(Snp[21], 0), (Snp[20], 1), (Snp[26], 1)],
        [(Snp[70], 0), (Snp[38], 1)],
        [(Snp[1], 1), (Snp[8], 0)],
        [(Snp[8], 1), (Snp[9], 0)],
        [(Snp[10], 0), (Snp[9], 1)],
        [(Snp[19], 0), (Snp[18], 1)],
    ]

    return connections


def simple_cnxs(f_n: int):
    freq = rf.Frequency(start=1, stop=2, npoints=f_n, unit="GHz")
    line = rf.media.DefinedGammaZ0(frequency=freq, z0=50)

    # network A
    ntwkA = rf.Network(name="a")
    ntwkA.frequency = freq
    ntwkA_z0 = (60, 70, 80, 90, 100)
    ntwkA.z0 = [ntwkA_z0] * f_n
    ntwkA.s = (
        np.random.default_rng()
        .random(f_n * len(ntwkA_z0) ** 2)
        .reshape(f_n, len(ntwkA_z0), len(ntwkA_z0))
    )

    # network B
    ntwkB = rf.Network(name="b")
    ntwkB.frequency = freq
    ntwkB_z0 = (10, 20, 30)
    ntwkB.z0 = [ntwkB_z0] * f_n
    ntwkB.s = (
        np.random.default_rng()
        .random(f_n * len(ntwkB_z0) ** 2)
        .reshape(f_n, len(ntwkB_z0), len(ntwkB_z0))
    )

    # Construct the connection
    port1 = rf.Circuit.Port(frequency=freq, name="port1", z0=50)
    port2 = rf.Circuit.Port(frequency=freq, name="port2", z0=50)
    gnd = rf.Circuit.Ground(frequency=freq, name="gnd", z0=50)
    open_port = rf.Circuit.Open(frequency=freq, name="open_port", z0=50)
    ind_shunt = line.inductor(50e-12, name="ind_shunt")
    cap_series = line.capacitor(30e-9, name="cap_series")
    port3 = rf.Circuit.Port(frequency=freq, name="port3", z0=50)
    port4 = rf.Circuit.Port(frequency=freq, name="port4", z0=50)

    connections = [
        [(ntwkB, 1), (port3, 0)],
        [(ntwkA, 3), (open_port, 0)],
        [(ntwkA, 0), (cap_series, 0)],
        [(port1, 0), (ind_shunt, 0), (cap_series, 1)],
        [(gnd, 0), (ind_shunt, 1), (ntwkA, 2)],
        [(ntwkA, 1), (port2, 0)],
        [(ntwkA, 4), (ntwkB, 0)],
        [(ntwkB, 2), (port4, 0)],
    ]

    return connections


def easy_cnxs(f_n: int):
    freq = rf.Frequency(start=1, stop=2, npoints=f_n, unit="GHz")
    line = rf.media.DefinedGammaZ0(frequency=freq, z0=50)

    # network A
    ntwkA = rf.Network(name="a")
    ntwkA.frequency = freq
    ntwkA_z0 = (60, 70, 80, 90, 100)
    ntwkA.z0 = [ntwkA_z0] * f_n
    ntwkA.s = (
        np.random.default_rng()
        .random(f_n * len(ntwkA_z0) ** 2)
        .reshape(f_n, len(ntwkA_z0), len(ntwkA_z0))
    )

    # Construct the connection
    port1 = rf.Circuit.Port(frequency=freq, name="port1", z0=50)
    port2 = rf.Circuit.Port(frequency=freq, name="port2", z0=50)
    port3 = rf.Circuit.Port(frequency=freq, name="port3", z0=50)
    port4 = rf.Circuit.Port(frequency=freq, name="port4", z0=50)
    port5 = rf.Circuit.Port(frequency=freq, name="port5", z0=50)

    connections = [
        [(ntwkA, 0), (port1, 0)],
        [(ntwkA, 1), (port2, 0)],
        [(ntwkA, 2), (port3, 0)],
        [(ntwkA, 3), (port4, 0)],
        [(ntwkA, 4), (port5, 0)],
    ]

    return connections


def test_reduce_circuit(connections):
    """
    Test block method and speed comparison with Index method.
    """
    # Index method, that is calculate the whole external S matrix, and index the ports
    index_start = time()
    index_rst = None
    index_rst = rf.Circuit(connections, auto_reduce=True).s_external
    index_end = time()

    return ((index_end - index_start) * 1000.0,)


if __name__ == "__main__":
    print(f"skrf path: {rf.__file__}")
    freq_n = 1001
    fcalls = 100

    for key, func in (
        ("Easy", easy_cnxs),
        ("Simple", simple_cnxs),
        ("Complex", complex_cnxs),
        ("HFSS", HFSS_cnxs),
    ):
        connections = func(freq_n)

        opted_cnxs = connections

        rst = np.array([test_reduce_circuit(opted_cnxs) for _ in range(fcalls)])

        print(
            f"Total time in {key} example: {np.mean(rst[:,0]):.4f} ms, min: {np.min(rst[:,0]):.4f} ms, std: {np.std(rst[:,0]):.4f} ms"
        )

        std_rst = rf.Circuit(connections)
        opt_rst = rf.Circuit(opted_cnxs, auto_reduce=True)

        print(f"Standard ({std_rst.dim}) and Optimized ({opt_rst.dim}) results are equal: {np.allclose(std_rst.s_external, opt_rst.s_external)}\n")
```
The following is a performance comparison of the `reduce_circuit` function in the master and this PR.
```
(PyQt) ➜  scikit-rf git:(CircuitPerformance) ✗ # Without Reorder/Split GND.
(PyQt) ➜  scikit-rf git:(CircuitPerformance) ✗ python Benchmark.py         
skrf path: /home/lansus/WorkFile/Git/scikit-rf/skrf/__init__.py
Total time in Easy example: 6.8258 ms, min: 6.1545 ms, std: 1.4882 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 14.1385 ms, min: 13.3212 ms, std: 0.9232 ms
Standard (18) and Optimized (12) results are equal: True

Total time in Complex example: 188.9378 ms, min: 127.7640 ms, std: 129.4097 ms
Standard (80) and Optimized (34) results are equal: True

Total time in HFSS example: 189.9583 ms, min: 149.7459 ms, std: 38.9025 ms
Standard (47) and Optimized (25) results are equal: True

(PyQt) ➜  scikit-rf git:(CircuitPerformance) ✗ # Reorder & Split GND. 
(PyQt) ➜  scikit-rf git:(CircuitPerformance) ✗ python Benchmark.py   
skrf path: /home/lansus/WorkFile/Git/scikit-rf/skrf/__init__.py
Total time in Easy example: 6.7455 ms, min: 5.9245 ms, std: 1.2488 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 11.1722 ms, min: 10.6947 ms, std: 1.0175 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 100.6201 ms, min: 81.7459 ms, std: 10.5824 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 90.6674 ms, min: 89.1821 ms, std: 3.2638 ms
Standard (47) and Optimized (13) results are equal: True
```