@Asachoo 2024-06-06
Sure, I'd be happy to explain the specific functions of these parameters.

The purpose of `check_duplication` is clear from the code: it checks for any duplicate `Network` names in the `connections_list`. This parameter defaults to True to ensure that there are no duplicate `Network` names in the result.

The `split_ground` parameter is used to split n networks sharing a common ground into n separate connections. Here is an example to illustrate the role of `split_ground` in the `reduce_circuit` method:
```python
import numpy as np

import skrf as rf


def build_circuit(f_n: int):
    freq = rf.Frequency(start=1, stop=2, npoints=f_n, unit="GHz")

    hfss_ntwk = rf.Network(name="HFSS")
    hfss_ntwk.s = np.random.rand(f_n, 16, 16) + 1.0j * np.random.rand(f_n, 16, 16)
    hfss_ntwk.frequency = freq

    line = rf.media.DefinedGammaZ0(frequency=freq, z0=50)

    Port1 = rf.Circuit.Port(frequency=freq, name="Port1")
    Port2 = rf.Circuit.Port(frequency=freq, name="Port2")
    Port3 = rf.Circuit.Port(frequency=freq, name="Port3")

    Ind1 = line.inductor(name="Ind1", L=3.85e-9)
    Ind2 = line.inductor(name="Ind2", L=5.94e-9)
    Ind3 = line.inductor(name="Ind3", L=3.14e-9)

    Cap1 = line.capacitor(name="Cap1", C=4.2e-12)
    Cap2 = line.capacitor(name="Cap2", C=0.35e-12)

    Gnd = rf.Circuit.Ground(frequency=freq, name="Gnd")

    ntwks = [rf.Network(name=f"ntwk{i}") for i in range(9)]
    for i in range(9):
        ntwks[i].frequency = freq
        ntwks[i].s = np.random.rand(f_n, 2, 2) + 1.0j * np.random.rand(f_n, 2, 2)

    connections = [
        [(hfss_ntwk, 10), (ntwks[5], 0)],
        [(hfss_ntwk, 11), (ntwks[6], 0)],
        [(hfss_ntwk, 12), (ntwks[7], 0)],
        [(hfss_ntwk, 0), (Ind1, 0), (Cap1, 0)],
        [(hfss_ntwk, 5), (ntwks[2], 0)],
        [(hfss_ntwk, 9), (ntwks[4], 0)],
        [
            (Ind1, 1),
            (Cap2, 1),
            (Ind3, 1),
            (ntwks[0], 1),
            (ntwks[1], 1),
            (ntwks[2], 1),
            (ntwks[4], 1),
            (ntwks[5], 1),
            (ntwks[6], 1),
            (hfss_ntwk, 14),
            (ntwks[8], 1),
            (Gnd, 0),
        ],
        [(hfss_ntwk, 7), (ntwks[3], 0)],
        [(hfss_ntwk, 8), (ntwks[3], 1)],
        [(Cap1, 1), (Port1, 0)],
        [(hfss_ntwk, 1), (Cap2, 0), (Ind2, 0)],
        [(Ind2, 1), (Port2, 0)],
        [(hfss_ntwk, 2), (Ind3, 0), (Port3, 0)],
        [(hfss_ntwk, 3), (ntwks[0], 0)],
        [(hfss_ntwk, 4), (ntwks[1], 0)],
        [(hfss_ntwk, 13), (ntwks[7], 1)],
        [(hfss_ntwk, 15), (ntwks[8], 0)],
    ]

    return connections


if __name__ == "__main__":
    initial_cnx = build_circuit(101)
    reduced_cnx = rf.reduce_circuit(initial_cnx)
    reduced_sg_cnx = rf.reduce_circuit(initial_cnx, split_ground=True)

    initial_ckt = rf.Circuit(initial_cnx)
    reduced_ckt = rf.Circuit(reduced_cnx)
    reduced_sg_ckt = rf.Circuit(reduced_sg_cnx)

    print(f"Initial Connection/Circuit: Node number {len(initial_cnx)}, Circuit dimension {initial_ckt.dim}")
    print(f"Reduced Connection/Circuit: Node number {len(reduced_cnx)}, Circuit dimension {reduced_ckt.dim}, Result is right? {np.allclose(initial_ckt.s_external, reduced_ckt.s_external)}")
    print(f"Reduced Connection/Circuit with split ground: Node number {len(reduced_sg_cnx)}, Circuit dimension {reduced_sg_ckt.dim}, Result is right? {np.allclose(initial_ckt.s_external, reduced_sg_ckt.s_external)}")

    initial_ckt.plot_graph(network_labels=True)
    reduced_ckt.plot_graph(network_labels=True)
    reduced_sg_ckt.plot_graph(network_labels=True)
```
The outputs
```
Initial Connection/Circuit: Node number 17, Circuit dimension 47
Reduced Connection/Circuit: Node number 6, Circuit dimension 25, Result is right? True
Reduced Connection/Circuit with split ground: Node number 5, Circuit dimension 13, Result is right? True
```
And some result figure
Initial Circuit:
![Initial](https://github.com/scikit-rf/scikit-rf/assets/50561614/e433725b-79ec-4723-90f0-a61cc669ec0f)
Reduced Circuit without split ground:
![Reduce_Without_SG](https://github.com/scikit-rf/scikit-rf/assets/50561614/5bb773be-0ea8-4a38-8dba-8dc645edadc0)
Reduced Circuit with split ground:
![Reduce_With_SG](https://github.com/scikit-rf/scikit-rf/assets/50561614/91e64a18-3778-489f-b771-9b299e2af8f8)
I hope this clarifies the functions of these two parameters.