@Asachoo 2024-06-11
> Before merging this, could you add some documentation by expanding or creating a notebook in the tutorial part?

Certainly, I'll add some documentation to the tutorial section after finding a suitable example.