@Asachoo 2024-06-15
> > Before merging this, could you add some documentation by expanding or creating a notebook in the tutorial part?
> 
> Certainly, I'll add some documentation to the tutorial section after finding a suitable example.

I've prepared a case and updated the notebook. What do you think of this example and its accompanying explanations? @jhillairet